atualização da lógica para se adequar a nova regra de negócio

+++ src/main/java/com/example/api_ponto_de_venda/dto/ItemVendaResponseDto.java
@@ -1,8 +1,8 @@
 package com.example.api_ponto_de_venda.dto;
 
 public record ItemVendaResponseDto(
-    Long id,
-    String nomeProduto,
-    Long quantidade,
-    double precoUnitario
+        Long id,
+        String nomeProduto,
+        Integer quantidade,
+        double precoUnitario
 ) {}
